Implement PooledVideoFrame
Implement a variant of VideoFrame which is compatible with FramePools.
PooledVideoFrame wraps an existing VideoFrame type and forwards all of
the core functionality to the inner frame, but its Drop implementation
returns it to the FramePool (if there hasn't been a DRC) so that it can
be re-used.

FramePools are necessary not only to avoid expensive DMA buffer
allocations, but to cap the number of File Descriptors in use at any
given time. Without FramePools, some test vectors will cause ccdec to
allocate a very large number of DMA buffers and exhaust our FD limit.

Bug: 390439314
Test: video.PlatformDecoding.cros_codecs* on Brya and Corsola
Change-Id: Ic31e24bfbc5eefda650009a6b674b81a9ae4d36a