Implement V4l2 Stateless DRC
Prior to this CL, the V4L2 stateless decoder did not have support for
bitstreams that had resolution changes.

This change updates the v4l2 stateless decoder and backend and updates
the h264 and vp8 codecs to support bitstreams that have resolution
changes. When a resolution change is detected, the decoder will wait for
all frames to be flushed out and then will reset the queues to have
them handle the updated resolution.

Bug: 392174530
Test: Ran ccdec on Corsola against |vp80-03-segmentation-1425.ivf|
Test: Converted |smpte_bars_resolution_ladder.h264.mp4| to .h264 using
ffmepg and decoded that bitstream using ccdec successfully.
Test: Ran vp8_comprehensive and vp8_inter and h264_baseline tast tests
on a Corsola successfully.

Change-Id: If487200bb6d3a2e350bc86aaa69461269a4b8644