Fix drain() behavior
C2 supports a draining mode that doesn't shut down the component, so we
should too.

Also remove the in flight queue from the decoder, since it causes a
subtle memory leak in H264. H264 gives the decoder frames one NALU at a
time, but the in flight queue is only popped once per frame, so we will
accumulate PPS and SPS NALUs.

Bug: 397758611
Test: video.PlatformDecoding.cros_codecs_vaapi_h264_baseline and
video.PlatformEncoding.cros_codecs_vaapi_h264_180 on Brya

Change-Id: I9175f59ce00fe6504ca9be7cee6b2be812944869