Update ccdec to check frame count.
Prior to this CL, the cros-codec decoder would not check if the decoder
was able to produce all frames if the golden argument was specified.
Rather it would only check the output frames against the golden frames.

This CL checks that the decoder outputs all frames by comparing the
output frame count against the number of golden frames provided.

Bug: 399224523
Test: Running 00000527.ivf on Brya now fails as expected since it outputs
no frames.

Change-Id: Ib0eee26fb301499d02774186983f6a944d363a80