Consume entire bitstream chunk in decode
Previously we assumed that the StatelessDecoder would consume the entire
bitstream chunk given in a C2DecodeJob, because the demuxer already
chunked up the bitstream into either NALUs for H264 or frames for VP8
and VP9. Unfortunately, this breaks AV1, which uses the IVF container
format, but requires further demuxing at the OBU level.

Bug: 399453671
Test: video.PlatformDecoding.cros_codecs_vaapi_av1* and
video.PlatformDecoding.cros_codecs_vaapi_h264* on Brya

Change-Id: I932ab486cc6b4cfe02ab566aee4016ff996ac07f